Major changes in client, fixed setup
Seperate Client to Client and HTTPConnection
The later is put in a separate package kamaki.client.connection
It manages headers, params, method, url and data of a connection
but the actuall http connection is not implemented at this
abstract level
There is an HTTPRequest class extention of HTTPConnection in the
same package. It implements connection using requests package
so basicaly it replicates the connection procedure that was used
in kamaki so far.
Minor modifications were nessacary at every client lib package

Setup is now aware of gevent and commands package