Fix old livetest pithos error, new _encode method
Old livetest pithos has an error in test_account_post since pithos
server started checking if usernames to be set in a group existed.
This change affected live this specific live kamaki test that used
to test the set_account_group operation with fake user ids.

New _encode method introduced in clients.connection.kamakicon that
checkes if a value is in unicode and url-encodes it ancordingly.
This functionality was already implemented as raw code wherever
needed, so now it's cleaner