Make authentication lazy
Refs: #4300

Authentication (a.k.a. POST /tokens) is performed only if needed, therefore a
syntax query will not request for authentication.

A new method is introduced in kamaki.cli: init_chached_authenticator. This
method authenticated the resolved URL and TOKEN(s).

Until now, this functionality was part of the _init_session method located at
kamaki.cli entry point. Now, this method call has been pushed down.

In case of an interactive shell, the method is called during the shell setup
In case of a one_command, the method is called right before a new executable
command is initialized. This also means that a non-executable command name
space will not cause the execution of init_cached_authenticator.