Move default argument initilization to kamaki.cli
Default arguments where set in kamaki.cli.arguments using the
global variable "_arguments". This was a semantic mistake, a
pep8 violation and arguably bad style.  The "_config_arg"
variable was also global and auxiliary to "_arguments"
initialization.

The "_argument" and "_config_arg" global vars are removed from
kamaki.cli.arguments. The initialization is now an internal
proccess in the "main" function of kamaki.cli, which is the
method responsible for setting up all kamaki interfaces.