Add ignore-ssl and ca-certs options to CLI
Refs #54

Both options are set either in the config file or as runtime
arguments.

ignore-ssl option allows connections without checking SSL certificates.

The ca-certs option provides a CA file for SSL authentication.

Normally, if no cert files are provided or they are invalid, kamaki
attempts to connect anyway, and an SSL Error ensues.

If ignore-ssl is set to "on", kamaki connects to servers regardless of
the existence or validity of a cert file and all SSL errors are ignored.

To help package maintainers to provide the correct CA file for their platform,
the default path is set in "kamaki/defaults.py" as CACERTS_DEFAULT_PATH