Check server status before changing it
Fixes grnet/kamaki#57

This fix affects the CLI commands "server_start", "server_reboot"
and "server_shutdown" which are specified in "kamaki.cli.cmds.cyclades".
These are the operations that modify the status of a server.

The problem was caused when the target status was the same as the current.
In this case, the request for status modification would fail with a 400
error code response. At this point, kamaki couldn't have adequate information
to correctly evaluate the error.

This problem is solved by checking the current status before attempting to
modify it.  Also, the server ID is validated in advance, instead of waiting
for a 400 server error.