file: Fix uploading from Windows
Up until now, if a user uploads a directory recursively from
a Windows machine, then the following file structure will be
stored in pithos:

   D dir1/
 10B dir1/dir2/dir3/file
   D dir1\dir2/
   D dir1\dir2\dir3

Downloading this structure from a Linux machine is impossible
because the following directories are created locally:

  dir1
  dir1\dir2
  dir1\dir2\dir3

This way the file cannot be stored under dir1/dir2/dir3 which
does not exist.

Fix this by storing all paths with '/' as a separator in Pithos,
rather than '\'.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>