Raise SSL error on IOError due to SSL cred. files
The problem was noted in OS X: when an SSL credential file (e.g.,
the certificates bundle) does not exist or is not provided, the
ssl.wrap_socket method fails with an IOError, while in other
systems it fails with an SSLError.

In this commit, the IOError is examined and, if some conditions
hold (errno==2, no SSL credential files or files do not exist), an
SSLError is raised.