server create: Add key_name argument
Following the synnefo's keypairs extensions, we' ve enriched the
kamaki cli in order to accept `key_name` as an optional parameter.
This argument should be mapped to a keypair model on the database.
These models can be retrieved from the
`/cyclades/compute/v2.0/os-keypairs` endpoint. Feature work would
be an extension in kamaki that would list the keypair details.