Support multiple key names in "server create"
In client API, the "create_server" method accepts two similar
parameters: "key_names" and "key_name". The first is a list of key
names and works only if the server supports the SNF:key_names
extention. The second is a single key name and works as described
by OS.

In CLI, the "server create" command accepts multiple "--key-name"
arguments. The API method is called with the "key_names" parameter
even if only one key name is provided.